requirements.txt
# requirements.txt - Compatible with Python 3.12
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
exifread==3.0.0
stegano==0.11.5  # Updated version for Python 3.12
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
slowapi==0.1.8
pandas==2.1.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
cryptography==41.0.7
python-magic==0.4.27  # For better file type detection
numpy==1.26.0  # For ELA calculations
reportlab==4.0.5  # For PDF reports
jinja2==3.1.2